requirements.txt
streamlit>=1.25
numpy>=1.23
opencv-python-headless>=4.7
pillow>=9.5
scikit-image>=0.21
scipy>=1.11
plotly>=5.16
matplotlib>=3.7
scikit-learn>=1.3
streamlit-webrtc>=0.45
